+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @definitelytyped/eslint-plugin
 
+## 0.0.187
+
+### Patch Changes
+
+- @definitelytyped/utils@0.0.183
+
+## 0.0.186
+
+### Patch Changes
+
+- ad779df9: fix: position of reporting ESLint rule no-unnecessary-generics
+
 ## 0.0.185
 
 ### Patch Changes